requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-bootstrap4==0.0.7
django-js-asset==1.1.0
django-mptt==0.9.1
idna==2.8
oauthlib==2.1.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.1.0
urllib3==1.24.1
